Made ly-per float

+++ trade.py
@@ -44,7 +44,7 @@ def parse_command_line():
     parser.add_argument('--hops', metavar="<Hops>", help="Number of hops to run", type=int, default=2, required=False)
     parser.add_argument('--jumps', dest='maxJumps', metavar="<Jumps>", help="Maximum total jumps", type=int, default=None, required=False)
     parser.add_argument('--jumps-per', dest='maxJumpsPer', metavar="<Jumps>", help="Maximum jumps per hop", type=int, default=3, required=False)
-    parser.add_argument('--ly-per', dest='maxLyPer', metavar='<LY>', help="Maximum light years per jump", type=int, default=8, required=False)
+    parser.add_argument('--ly-per', dest='maxLyPer', metavar='<LY>', help="Maximum light years per jump", type=float, default=8, required=False)
     parser.add_argument('--capacity', metavar="<Capactiy>", help="Maximum capacity of cargo hold", type=int, default=4, required=False)
     parser.add_argument('--limit', help='Maximum units of any one cargo item to buy', type=int, default=0, required=False)
     parser.add_argument('--unique', help='Only visit each station once', default=False, required=False, action='store_true')
